Fix matchFilterTitle condition in selectors and Add setDataToStore.js to helpers
