+++ src/assets/styles/abstracts/_variables.scss
@@ -2,7 +2,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@200;400;700&display=swap');
 
 // COLORS
-$primary-color: rgb(253, 253, 168);
+$primary-color: #fdfda8;
 $secondary-color: rgb(255, 64, 239);
 $grey: #777;
 $light-grey-1: #ccc;
